more eval sample summary alongside eval sample
