+++ tests/log/test_log_eventdb_sync.py
@@ -3,13 +3,14 @@ from pathlib import Path
 
 import pytest
 
+from inspect_ai.log._log import EvalSampleSummary
 from inspect_ai.log._recorders.buffer.database import (
     SampleBufferDatabase,
     sync_to_filestore,
 )
 from inspect_ai.log._recorders.buffer.filestore import SampleBufferFilestore
 from inspect_ai.log._recorders.buffer.types import Samples
-from inspect_ai.log._recorders.types import EvalSampleSummary, SampleEvent
+from inspect_ai.log._recorders.types import SampleEvent
 from inspect_ai.log._transcript import InfoEvent
